Slight refactor for usability
